make setDNSCachingTime() take a long for consistency.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@20671 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/cURLManager.h
@@ -50,7 +50,7 @@ public:
   void setTimeCondition(timeCondition condition, time_t &t);
   void setInterface(const std::string interfaceIP);
   void setUserAgent(const std::string userAgent);
-  void setDNSCachingTime(int time);
+  void setDNSCachingTime(long time);
   void setDeleteOnDone();
 
   void addFormData(const char *key, const char *value);
